// frontend/src/app/learncoreservices/models/create-question-form.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateQuestionForm {
  categoryUuid: string;
  difficultyLevel: 'EASY' | 'MEDIUM' | 'HARD' | 'EXPERT';
  explanation?: string;
  questionText: string;
  questionType: 'MCQ' | 'TRUE_OR_FALSE' | 'STRUCTURAL' | 'ESSAY';

  /**
   * Optional JSON string for STRUCTURAL/ESSAY questions
   */
  writtenAnswerConfig?: string;
}