// frontend/src/app/learncoreservices/models/get-question-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WrittenAnswerConfigResponse } from '../models/written-answer-config-response';
export interface GetQuestionResponse {
  categoryName?: string;
  categoryUuid?: string;
  difficultyLevel?: 'EASY' | 'MEDIUM' | 'HARD' | 'EXPERT';
  explanation?: string;
  imageUrl?: string;
  questionText?: string;
  questionType?: 'MCQ' | 'TRUE_OR_FALSE' | 'STRUCTURAL' | 'ESSAY';
  uuid?: string;
  writtenAnswerConfig?: WrittenAnswerConfigResponse;
}