// frontend/src/app/learncoreservices/models/update-question-form.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateQuestionForm {
  categoryUuid: string;
  difficultyLevel: 'EASY' | 'MEDIUM' | 'HARD' | 'EXPERT';
  explanation?: string;
  questionText: string;
  questionType: 'MCQ' | 'TRUE_OR_FALSE' | 'STRUCTURAL' | 'ESSAY';

  /**
   * Optional JSON string for STRUCTURAL/ESSAY questions
   */
  writtenAnswerConfig?: string;
}